- modularized the audio panel
