+++ app/main_window.py
@@ -1,6 +1,6 @@
 import customtkinter
 from audio_manager import get_config_file, construct_audio_loop
-from component import Console
+from component import Console, AudioPanel
 
 customtkinter.set_appearance_mode("System")  # Modes: "System" (standard), "Dark", "Light"
 customtkinter.set_default_color_theme("blue")  # Themes: "blue" (standard), "green", "dark-blue"
@@ -52,28 +52,9 @@ class App(customtkinter.CTk):
 
 
         #region 选择音频菜单GUI
-        self.audio_frame = customtkinter.CTkFrame(self)
+        self.audio_frame = AudioPanel(self, start_engine=self.start_engine)
         self.audio_frame.grid(row=0, column=1, rowspan=2, padx=(20, 20), pady=(20, 10), sticky="nsew")
         self.audio_frame.grid_columnconfigure(0, weight=1)  # configure grid of individual tabs
-
-        self.label_audio_group = customtkinter.CTkLabel(master=self.audio_frame, text="音频控制台", font=FONT_TITLE)
-        self.label_audio_group.grid(row=0, column=0, padx=10, pady=5)
-
-        customtkinter.CTkLabel(self.audio_frame, text="请选择机经序号" ).grid(row=1, column=0, padx=20)
-
-        self.combobox_1 = self._init_audio_option_menu()
-        self.combobox_1.grid(row=2, column=0, padx=20, pady=(5, 5))
-
-        self.string_input_button = customtkinter.CTkButton(self.audio_frame, text="选择",
-                                                           command=self.start_engine)
-        self.string_input_button.grid(row=3, column=0, padx=20, pady=(10, 10))
-
-
-        self.audio_indicator = customtkinter.CTkLabel(self.audio_frame, text="-- 未选择音频 --")
-        self.audio_indicator.grid(row=4, column=0, padx=20, pady=(10, 10))
-
-        self.space_holder_1 =  customtkinter.CTkLabel(self.audio_frame, text=" ")
-        self.space_holder_1.grid(row=5, column=0, padx=20, pady=(10, 10))
         #endregion
 
 
@@ -121,19 +102,6 @@ class App(customtkinter.CTk):
         self.next_button.grid(row=2, column=2, padx=(10, 20), pady=(10, 10))
         #endregion
 
-
-    #region 初始化GUI相关
-    def _init_audio_option_menu(self):
-        config_data = get_config_file()
-        options = [str(x["id"]) + " " + x["file_name"] for x in config_data] if config_data else []
-        combobox = customtkinter.CTkOptionMenu(self.audio_frame, dynamic_resizing=False, 
-                                           values=options)
-        if options is not None and len(options) > 0:
-            combobox.set(options[0])
-        else:
-            combobox.set("-- Empty --")
-        return combobox
-    #endregion
 
     #region 播放功能相关
     def start_engine(self):
