+++ component/__init__.py
@@ -1,1 +1,2 @@
 from .console import Console
+from .audio_panel import AudioPanel
